Revert "Merge pull request #122 from nelz07/dev.V3"
This reverts commit 75fe08fa7b1eaf8f39a6432be6e46e3803265e71, reversing
changes made to 97de9b5b994e80ee7a00ded758356e7b42389d86.

+++ app/DepositPayment.php
@@ -38,7 +38,6 @@ class DepositPayment extends Model
 
     public function revert($user_id){
         $data = $this->revertData($user_id);
-        
         $this->update([
             'reverted'=>true,
             'reverted_by'=>$user_id
@@ -48,7 +47,7 @@ class DepositPayment extends Model
 
     public function revertData($user_id){
         
-        $data['amount'] = (double) $this->amount;
+        $data['amount'] = (int) $this->amount;
         $data['transaction_number'] = 'RD'.str_replace('.','',microtime(true));
         $data['payment_method_id'] = (int) $this->payment_method_id;
         $data['notes'] = 'Revertion for ' . $this->transaction_number;
